+++ internal/cmd/branch.go
@@ -65,7 +65,7 @@ func executeBranch(cliConfig cliconfig.CliConfig) error {
 	if err != nil || exit {
 		return err
 	}
-	entries := NewSwitchBranchEntries(SwitchBranchArgs{
+	entries := dialog.NewSwitchBranchEntries(dialog.NewSwitchBranchEntriesArgs{
 		BranchInfos:       data.branchInfos,
 		BranchTypes:       []configdomain.BranchType{},
 		BranchesAndTypes:  data.branchesAndTypes,
