+++ internal/cmd/switch_test.go
@@ -5,7 +5,6 @@ import (
 	"testing"
 
 	"github.com/git-town/git-town/v21/internal/cli/dialog"
-	"github.com/git-town/git-town/v21/internal/cmd"
 	"github.com/git-town/git-town/v21/internal/config/configdomain"
 	"github.com/git-town/git-town/v21/internal/git/gitdomain"
 	"github.com/git-town/git-town/v21/internal/regexes"
@@ -40,7 +39,7 @@ func TestSwitchBranch(t *testing.T) {
 				branchesAndTypes := configdomain.BranchesAndTypes{}
 				unknownBranchType := configdomain.UnknownBranchType(configdomain.BranchTypeFeatureBranch)
 				regexes := []*regexp.Regexp{}
-				have := cmd.NewSwitchBranchEntries(cmd.SwitchBranchArgs{
+				have := dialog.NewSwitchBranchEntries(dialog.NewSwitchBranchEntriesArgs{
 					BranchInfos:       branchInfos,
 					BranchTypes:       branchTypes,
 					BranchesAndTypes:  branchesAndTypes,
@@ -71,7 +70,7 @@ func TestSwitchBranch(t *testing.T) {
 				branchesAndTypes := configdomain.BranchesAndTypes{}
 				unknownBranchType := configdomain.UnknownBranchType(configdomain.BranchTypeFeatureBranch)
 				regexes := []*regexp.Regexp{}
-				have := cmd.NewSwitchBranchEntries(cmd.SwitchBranchArgs{
+				have := dialog.NewSwitchBranchEntries(dialog.NewSwitchBranchEntriesArgs{
 					BranchInfos:       branchInfos,
 					BranchTypes:       branchTypes,
 					BranchesAndTypes:  branchesAndTypes,
@@ -106,7 +105,7 @@ func TestSwitchBranch(t *testing.T) {
 			branchesAndTypes := configdomain.BranchesAndTypes{}
 			unknownBranchType := configdomain.UnknownBranchType(configdomain.BranchTypeFeatureBranch)
 			regexes := []*regexp.Regexp{}
-			have := cmd.NewSwitchBranchEntries(cmd.SwitchBranchArgs{
+			have := dialog.NewSwitchBranchEntries(dialog.NewSwitchBranchEntriesArgs{
 				BranchInfos:       branchInfos,
 				BranchTypes:       branchTypes,
 				BranchesAndTypes:  branchesAndTypes,
@@ -140,7 +139,7 @@ func TestSwitchBranch(t *testing.T) {
 				branchesAndTypes := configdomain.BranchesAndTypes{}
 				unknownBranchType := configdomain.UnknownBranchType(configdomain.BranchTypeFeatureBranch)
 				regexes := []*regexp.Regexp{}
-				have := cmd.NewSwitchBranchEntries(cmd.SwitchBranchArgs{
+				have := dialog.NewSwitchBranchEntries(dialog.NewSwitchBranchEntriesArgs{
 					BranchInfos:       branchInfos,
 					BranchTypes:       branchTypes,
 					BranchesAndTypes:  branchesAndTypes,
@@ -174,7 +173,7 @@ func TestSwitchBranch(t *testing.T) {
 				branchesAndTypes := configdomain.BranchesAndTypes{}
 				unknownBranchType := configdomain.UnknownBranchType(configdomain.BranchTypeFeatureBranch)
 				regexes := []*regexp.Regexp{}
-				have := cmd.NewSwitchBranchEntries(cmd.SwitchBranchArgs{
+				have := dialog.NewSwitchBranchEntries(dialog.NewSwitchBranchEntriesArgs{
 					BranchInfos:       branchInfos,
 					BranchTypes:       branchTypes,
 					BranchesAndTypes:  branchesAndTypes,
@@ -205,7 +204,7 @@ func TestSwitchBranch(t *testing.T) {
 				branchesAndTypes := configdomain.BranchesAndTypes{}
 				unknownBranchType := configdomain.UnknownBranchType(configdomain.BranchTypeFeatureBranch)
 				regexes := []*regexp.Regexp{}
-				have := cmd.NewSwitchBranchEntries(cmd.SwitchBranchArgs{
+				have := dialog.NewSwitchBranchEntries(dialog.NewSwitchBranchEntriesArgs{
 					BranchInfos:       branchInfos,
 					BranchTypes:       branchTypes,
 					BranchesAndTypes:  branchesAndTypes,
@@ -247,7 +246,7 @@ func TestSwitchBranch(t *testing.T) {
 				}
 				unknownBranchType := configdomain.UnknownBranchType(configdomain.BranchTypeFeatureBranch)
 				regexes := []*regexp.Regexp{}
-				have := cmd.NewSwitchBranchEntries(cmd.SwitchBranchArgs{
+				have := dialog.NewSwitchBranchEntries(dialog.NewSwitchBranchEntriesArgs{
 					BranchInfos:       branchInfos,
 					BranchTypes:       branchTypes,
 					BranchesAndTypes:  branchesAndTypes,
@@ -287,7 +286,7 @@ func TestSwitchBranch(t *testing.T) {
 				}
 				unknownBranchType := configdomain.UnknownBranchType(configdomain.BranchTypeFeatureBranch)
 				regexes := []*regexp.Regexp{}
-				have := cmd.NewSwitchBranchEntries(cmd.SwitchBranchArgs{
+				have := dialog.NewSwitchBranchEntries(dialog.NewSwitchBranchEntriesArgs{
 					BranchInfos:       branchInfos,
 					BranchTypes:       branchTypes,
 					BranchesAndTypes:  branchesAndTypes,
@@ -324,7 +323,7 @@ func TestSwitchBranch(t *testing.T) {
 				unknownBranchType := configdomain.UnknownBranchType(configdomain.BranchTypeFeatureBranch)
 				regexes, err := regexes.NewRegexes([]string{})
 				must.NoError(t, err)
-				have := cmd.NewSwitchBranchEntries(cmd.SwitchBranchArgs{
+				have := dialog.NewSwitchBranchEntries(dialog.NewSwitchBranchEntriesArgs{
 					BranchInfos:       branchInfos,
 					BranchTypes:       branchTypes,
 					BranchesAndTypes:  branchesAndTypes,
@@ -359,7 +358,7 @@ func TestSwitchBranch(t *testing.T) {
 				unknownBranchType := configdomain.UnknownBranchType(configdomain.BranchTypeFeatureBranch)
 				regexes, err := regexes.NewRegexes([]string{"observed-"})
 				must.NoError(t, err)
-				have := cmd.NewSwitchBranchEntries(cmd.SwitchBranchArgs{
+				have := dialog.NewSwitchBranchEntries(dialog.NewSwitchBranchEntriesArgs{
 					BranchInfos:       branchInfos,
 					BranchTypes:       branchTypes,
 					BranchesAndTypes:  branchesAndTypes,
@@ -391,7 +390,7 @@ func TestSwitchBranch(t *testing.T) {
 				unknownBranchType := configdomain.UnknownBranchType(configdomain.BranchTypeFeatureBranch)
 				regexes, err := regexes.NewRegexes([]string{"observed-", "main"})
 				must.NoError(t, err)
-				have := cmd.NewSwitchBranchEntries(cmd.SwitchBranchArgs{
+				have := dialog.NewSwitchBranchEntries(dialog.NewSwitchBranchEntriesArgs{
 					BranchInfos:       branchInfos,
 					BranchTypes:       branchTypes,
 					BranchesAndTypes:  branchesAndTypes,
